The lat long from the map now get handed to the input elements.
