+++ src/aims_ui/templates/aims_ui_pages/a_single_matches/radiussearch.html
@@ -83,7 +83,7 @@
 
 <!-- Load Leaflet JS before page specific scripts that use the interactive map that relies on leaflet -->
 <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js" integrity="sha256-20nQCchB9co0qIjJZRGuk2/Z9VM+kNiyxNV1lvTlZBo=" crossorigin=""></script>
-<script type=module src="static/js/a_single_matches/radiussearch.mjs"></script>
+<script type=module src="static/js/a_single_matches/radiussearch/init.mjs"></script>
 <!-- We do not need custom column widths here as this page's layout is handled by CSS -->
 
 <script type=module src="static/js/e_all_pages/all_pages_last.mjs"></script>
